+++ testing/alerting-gen/pkg/execute/config.go
@@ -1,4 +1,4 @@
-package config
+package execute
 
 import (
 	"errors"
