+++ testing/alerting-gen/pkg/execute/run.go
@@ -11,7 +11,6 @@ import (
 
 	kitlog "github.com/go-kit/log"
 	"github.com/go-kit/log/level"
-	"github.com/grafana/alerting/testing/alerting-gen/pkg/config"
 	gen "github.com/grafana/alerting/testing/alerting-gen/pkg/generate"
 	api "github.com/grafana/grafana-openapi-client-go/client"
 	"github.com/grafana/grafana-openapi-client-go/client/folders"
@@ -19,7 +18,7 @@ import (
 	"github.com/grafana/grafana-openapi-client-go/models"
 )
 
-func Run(cfg config.Config, debug bool) ([]*models.AlertRuleGroup, error) {
+func Run(cfg Config, debug bool) ([]*models.AlertRuleGroup, error) {
 	if err := cfg.Validate(); err != nil {
 		return nil, err
 	}
@@ -114,7 +113,7 @@ func Run(cfg config.Config, debug bool) ([]*models.AlertRuleGroup, error) {
 
 // sendViaProvisioning maps the generated export groups into provisioned group payloads
 // and pushes them to Grafana using the provisioning API.
-func sendViaProvisioning(cfg config.Config, groups []*models.AlertRuleGroup, logger kitlog.Logger) error {
+func sendViaProvisioning(cfg Config, groups []*models.AlertRuleGroup, logger kitlog.Logger) error {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
@@ -221,7 +220,7 @@ func newGrafanaClient(baseURL, username, password, token string, orgID int64) (*
 }
 
 // nukeFolders deletes all alerting-gen created folders.
-func nukeFolders(cfg config.Config, logger kitlog.Logger) error {
+func nukeFolders(cfg Config, logger kitlog.Logger) error {
 	cli, err := newGrafanaClient(cfg.GrafanaURL, cfg.Username, cfg.Password, cfg.Token, cfg.OrgID)
 	if err != nil {
 		return err
@@ -267,7 +266,7 @@ func nukeFolders(cfg config.Config, logger kitlog.Logger) error {
 }
 
 // createFolders creates N folders in Grafana and returns their UIDs.
-func createFolders(cfg config.Config, numFolders int, seed int64, logger kitlog.Logger) ([]string, error) {
+func createFolders(cfg Config, numFolders int, seed int64, logger kitlog.Logger) ([]string, error) {
 	cli, err := newGrafanaClient(cfg.GrafanaURL, cfg.Username, cfg.Password, cfg.Token, cfg.OrgID)
 	if err != nil {
 		return nil, err
